Renamed borda client package
